fix: autofocus on themedModal
